% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annot_counts.R
\name{gene_annot_counts}
\alias{gene_annot_counts}
\title{gene annotation counts}
\usage{
gene_annot_counts(
  dt_gen,
  dt_snpgene,
  keep_indiv = NULL,
  extract_SNP = NULL,
  filter_gene = NULL
)
}
\arguments{
\item{dt_gen}{recoded genetic data from PLINK}

\item{dt_snpgene}{with SNP and GENE as column names}
}
\value{
Returns an object of data.table class as an output with allelic gene counts within each sample where each row corresponds to gene and column to individual IDs from column second. The first column contains gene names.
}
\description{
Inputs needed are recoded genetic data formatted in PLINK format and SNP-gene annotation data . The first six columns of the input genetic data follow standard PLINK .raw formt. Column names as FID, IID, PAT, MAT, SEX and PHENOTYPE followed by SNP information as recoded by the PLINK software. SNP-gene data has two columns: GENE and SNP names. The function returns allelic counts per gene per sample (where each row represents a gene and each column represents an individual starting with the second column where first column contains gene information).
}
\examples{

#Package provides sample data that are loaded with package loading. 

data(recodedgen) #PLINK raw formatted data of 10 individiduals with 10 SNPs

data(snpgene) #SNP and its respective GENE annotated. 
#Here 10 SNPs are shown annotated in five genes. 
#A SNP can be annotated in multiple genes. 

gene_annot_counts(recodedgen,snpgene) #run the function

#end not RUN

}
\author{
Sanjeev Sariya
}
