% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plinkSNP_index_geneboundaries.R
\name{get_plinkSNP_index_geneboundaries}
\alias{get_plinkSNP_index_geneboundaries}
\title{Provide SNP index using genetic boundaries and PLINK bim file}
\usage{
get_plinkSNP_index_geneboundaries(plinkbim_file,genecoord_dt)
}
\arguments{
\item{genecoord_dt}{a dataframe for gene boundaries with CHR START END GENE as column names. Where CHR should be integer 1-22. START and END column should be integer. GENE column contains gene names}

\item{plink_file}{plink .bed file with path}
}
\value{
a vector of integer that represents SNP index that are found within the gene boundaries and SNP bim file based on the interection. Returned values are unique if found, otherwise NULL is returned.
}
\description{
Function returns a vector of unique integer values for SNP position/index using genetic boundaries and plink bim file
}
\details{
PLINK (.bim) file will be read
}
\author{
Sanjeev Sariya
}
