% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_count_snppos.R
\name{plink_count_snppos}
\alias{plink_count_snppos}
\title{plink gene counts using SNP coords}
\usage{
plink_count_snppos(
  plink_file,
  genecoord_dt,
  snp_pos_dt,
  snp_index,
  individuals_index
)
}
\arguments{
\item{plink_file}{plink .bed file with path}

\item{genecoord_dt}{a dataframe for gene boundaries with CHR START END GENE as column names. Where CHR should be integer 1-22. START and END column should be integer. GENE column contains gene names}

\item{snp_pos_dt}{a dataframe for SNP information with SNP BP as column names.}

\item{snp_index}{a vector of integer that specifies SNPs to read.}

\item{individuals_index}{a vector of integer that specifies individuals to select.}
}
\description{
Function returns matrix with allelic counts per gene per individual for SNP and gene coordinates as inputs for PLINK (.bed) file format
}
\details{
PLINK (.bed) file will be read
}
