% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recodedgen}
\alias{recodedgen}
\title{data set prepared for the package
recodedgen is genetic data with 10 rows and 16 columns.
Where the first 6 columns are FID, IID, PAT, MAT, SEX and PHENOTYPE which are inherent from the
PLINK output. Columns followed by PHENOTYPE are SNP names followed by _A, or _C
or _T or _G that is allele. Each SNP column may have 0, 1, 2 or NA value}
\format{
An object of class \code{data.frame} with 10 rows and 16 columns.
}
\usage{
recodedgen
}
\description{
data set prepared for the package
recodedgen is genetic data with 10 rows and 16 columns.
Where the first 6 columns are FID, IID, PAT, MAT, SEX and PHENOTYPE which are inherent from the
PLINK output. Columns followed by PHENOTYPE are SNP names followed by _A, or _C
or _T or _G that is allele. Each SNP column may have 0, 1, 2 or NA value
}
\keyword{datasets}
