\docType{data}
\name{sample_chr21_vcf}
\alias{sample_chr21_vcf}
\title{sample_chr21_vcf}
\format{
VCF sample
}
\usage{
sample_chr21_vcf
}
\description{
VCF data hg19 build 
}
\details{
sample_chr21_vcf is sample VCF file provided for CHR21 for of ten individuals 
contains 128 SNPs
}
\keyword{datasets}
