% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpgene}
\alias{snpgene}
\title{snpgene}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\usage{
snpgene
}
\description{
sample SNP-gene annotation data
}
\details{
snpgene is sample SNP-Gene annotation data provided with the GARCOM package. It has 10 rows and 2 columns.
Column names are GENE and SNP, where GENE column contains GENE names and SNP column
contains SNP name that is annotated with the GENE
}
\keyword{datasets}
