% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snppos}
\alias{snppos}
\title{snppos SNP-BP data with 2 columns and 10 rows provided with the GARCOM package. Column names are SNP and BP.
where SNP columns contains SNP names BP column contains position of the SNP}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\usage{
snppos
}
\description{
snppos SNP-BP data with 2 columns and 10 rows provided with the GARCOM package. Column names are SNP and BP.
where SNP columns contains SNP names BP column contains position of the SNP
}
\keyword{datasets}
