% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_counts_annot.R
\name{vcf_counts_annot}
\alias{vcf_counts_annot}
\title{gene annotation counts using VCF data}
\usage{
vcf_counts_annot(vcf_data, df_snpgene)
}
\arguments{
\item{vcf_data}{an object of vcfR class}

\item{df_snpgene}{a data frame that contains SNP and annotated gene with SNP and gene as column name}
}
\value{
Returns an matrix of data.table class as an output with allelic (reference) gene counts within each sample where each row corresponds to gene and column to individual IDs from column second. The first column contains gene names.
}
\description{
Function returns a matrix with allelic (reference) counts per gene per individual for SNP-gene annotation
}
\details{
Inputs needed are a vcf data and a data frame of SNP-gene annotation. The function returns a matrix of allelic counts (reference) per gene per sample (where each row represents a gene and each column represents an individual starting with the second column where first column contains gene information).
}
\examples{
\dontrun{
vcf_counts_annot(vcf,df_snpgene_test)
}
}
\author{
Sanjeev Sariya
}
