% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_counts_withSNPpos.R
\name{vcf_counts_SNP_genecoords}
\alias{vcf_counts_SNP_genecoords}
\title{VCF gene position counts}
\usage{
vcf_counts_SNP_genecoords(vcf_data, df_snppos, df_genecoords)
}
\arguments{
\item{vcf_data}{an object of vcfR class}

\item{df_snppos}{a dataframe for SNP information with SNP BP as column names.}

\item{df_genecoords}{a dataframe for gene boundaries with CHR START END GENE as column names. Where CHR should be integer 1-22. START and END column should be integer. GENE column contains gene names}
}
\value{
Returns an matrix of data.table class as an output with allelic (reference) gene counts within each sample where each row corresponds to gene and column to individual IDs from column second. The first column contains gene names.
}
\description{
Function returns a matrix with allelic counts per gene per individual for SNP and gene coordinates as inputs
}
\examples{
\dontrun{
vcf_counts_SNP_genecoords(vcf_data_test,df_snppos_test,df_genecoords_test)
}

}
\author{
Sanjeev Sariya
}
