% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annot_counts.R
\name{gene_annot_counts}
\alias{gene_annot_counts}
\title{gene annotation counts}
\usage{
gene_annot_counts(
  dt_gen,
  dt_snpgene,
  keep_indiv = NULL,
  extract_SNP = NULL,
  filter_gene = NULL,
  impute_missing = FALSE,
  impute_method = "mean"
)
}
\arguments{
\item{dt_gen}{recoded genetic data from PLINK}

\item{dt_snpgene}{with SNP and GENE as column names}

\item{keep_indiv}{individuals to keep. mutation counts will be provided for individuals provided in the list only. Default all individuals are used.}

\item{extract_SNP}{SNPs to extract. mutation counts will be provided for SNPs provided in the list only. Default all SNPs are used.}

\item{filter_gene}{Genes to filter in. mutation counts will be provided for genes provided in the list only. Default all genes are used.}

\item{impute_missing}{default is FALSE.}

\item{impute_method}{default method used to impute missing values is mean. mean and median two methods are supported. Function accepts method in quotes: "mean" or "median"}
}
\value{
Returns an object of data.table class as an output with allelic gene counts within each sample where each row corresponds to gene and column to individual IDs from column second. The first column contains gene names.
}
\description{
Inputs needed are recoded genetic data formatted in PLINK format and SNP-gene annotation data . The first six columns of the input genetic data follow standard PLINK .raw formt. Column names as FID, IID, PAT, MAT, SEX and PHENOTYPE followed by SNP information as recoded by the PLINK software. SNP-gene data has two columns: GENE and SNP names. The function returns allelic counts per gene per sample (where each row represents a gene and each column represents an individual starting with the second column where first column contains gene information).
}
\examples{

#Package provides sample data that are loaded with package loading. 

data(recodedgen) #PLINK raw formatted data of 10 individiduals with 10 SNPs

data(snpgene) #SNP and its respective GENE annotated. 
#Here 10 SNPs are shown annotated in five genes. 
#A SNP can be annotated in multiple genes. 

gene_annot_counts(recodedgen,snpgene) #run the function

#subset Genes
gene_annot_counts(recodedgen,snpgene,filter_gene=c("GENE1","GENE2"))

#Subset individuals
gene_annot_counts(recodedgen, snpgene,keep_indiv=c("IID_sample1","IID_sample8"))

#subset with genes and samples
gene_annot_counts(recodedgen,snpgene,filter_gene=c("GENE1","GENE2"),
keep_indiv=c("IID_sample1","IID_sample8"))

#impute missing using default method. 
#Data are rounded to the two deimal places. 0.1234 will become 0.12.
gene_annot_counts(recodedgen,snpgene,impute_missing=TRUE)

#Subset on individuals and impute for missing values. Default as mean
gene_annot_counts(recodedgen,snpgene,impute_missing=TRUE,
keep_indiv=c("IID_sample1","IID_sample2","IID_sample10"))

#impute using median method
gene_annot_counts(recodedgen,snpgene,impute_missing=TRUE,impute_method="median")

#end not RUN

}
\author{
Sanjeev Sariya
}
