% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genecoord}
\alias{genecoord}
\title{genecoord is Gene coordinate data with 3 columns and 5 rows provided with the GARCOM package. Column names are GENE, START and END
where GENE column contains GENE name, START and END have information
for gene start and end coordinates respectively.}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\usage{
genecoord
}
\description{
genecoord is Gene coordinate data with 3 columns and 5 rows provided with the GARCOM package. Column names are GENE, START and END
where GENE column contains GENE name, START and END have information
for gene start and end coordinates respectively.
}
\keyword{datasets}
