% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genecoord}
\alias{genecoord}
\title{genecoord}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\usage{
genecoord
}
\description{
sample data for gene base pair boundaries
}
\details{
genecoord is example data provided with the GARCOM package. It has 3 columns and 5 rows.
Column names are GENE, START and END where GENE column contains gene name, START and END indicate
start BP and end BP respectively.
}
\keyword{datasets}
