% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snppos}
\alias{snppos}
\title{snppos}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\usage{
snppos
}
\description{
sample data for SNP coordinates
}
\details{
snppos is sample SNP-BP data provided with the GARCOM package. It has 2 columns and 10 rows. Column names are SNP and BP.
where SNP columns contains SNP names BP column contains position of the SNP
}
\keyword{datasets}
